refactor: errorElement를 사용해서 경로관련 오류를 처리할 수 있도록 변경
