+++ src/App.tsx
@@ -1,17 +1,15 @@
-import { BrowserRouter } from 'react-router-dom';
-import Router from './Router';
+import { RouterProvider } from 'react-router-dom';
+import router from './Router';
 import ThemeSwitcher from './components/ThemeSwitcher';
 import { BookStoreThemeProvider } from './context/themeContext';
 
 function App() {
   return (
     <div className="App">
-      <BrowserRouter>
-        <BookStoreThemeProvider>
-          <ThemeSwitcher />
-          <Router />
-        </BookStoreThemeProvider>
-      </BrowserRouter>
+      <BookStoreThemeProvider>
+        <ThemeSwitcher />
+        <RouterProvider router={router} />
+      </BookStoreThemeProvider>
     </div>
   );
 }
